test_security.rs enabled
