--- dash-spv-ffi/tests/security/mod.rs
@@ -1,1 +1,0 @@
-mod test_security;
